+++ src/main.rs
@@ -17,8 +17,8 @@ git_testament!(TESTAMENT);
 #[derive(Subcommand)]
 #[allow(clippy::large_enum_variant)]
 enum Commands {
-    /// Checks a WDL file (or a directory containing WDL files) and reports
-    /// diagnostics.
+    /// Checks a WDL document (or a directory containing WDL documents) and
+    /// reports diagnostics.
     Check(commands::check::CheckArgs),
 
     /// Lints Workflow Description Language files.
@@ -29,6 +29,10 @@ enum Commands {
 
     /// Runs the analyzer LSP server.
     Analyzer(commands::analyzer::AnalyzerArgs),
+
+    /// Formats a WDL document.
+    #[clap(alias = "fmt")]
+    Format(commands::format::FormatArgs),
 }
 
 #[derive(Parser)]
@@ -58,6 +62,7 @@ pub async fn inner() -> anyhow::Result<()> {
         Commands::Lint(args) => commands::check::lint(args).await,
         Commands::Explain(args) => commands::explain::explain(args),
         Commands::Analyzer(args) => commands::analyzer::analyzer(args).await,
+        Commands::Format(args) => commands::format::format(args),
     }
 }